sitemap: bring SEO back with versions in URLs
Apparently the use of `grass-stable/` and `grass-devel/` penalizes the new GRASS GIS 8 versions over G78 in common search engines.
Reason may be that they are only redirects which cannot be easily ranked.

This PR brings back the GRASS GIS version in the manual URLs (followup to #230) and to be considered in #318.