Mentoring program announcement (#393)
News item about launching NSF funded mentoring program.

Info regarding NSF Logos and Usage Standards:
NSF's logo can be used by recipients of NSF support for the sole purpose of acknowledging that support. NSF's logo can be used to link to an NSF website or acknowledge NSF assistance or affiliation. NSF's logo cannot be used in a manner that falsely implies employment by or affiliation with NSF. NSF's logo cannot be used to imply or endorse a product or service. No explicit permission is necessary to use the NSF logo. The NSF logo is the sole visual identifier of the U.S. National Science Foundation and embodies a set of values that is applied to everything we create and communicate. Source: https://www.nsf.gov/policies/logos.jsp (2023-10-09) 

---------

Co-authored-by: Veronica Andreo <veroandreo@gmail.com>