Use inverted white logo for index page
This uses a single color inverted logo in white (the grass plant is transparent, its background is white, everything else is white too).

The mission text is in bold font used for headings and is white and little smaller.

The grass photo in the background is not covered in gradient to white, but with transparent dark green overlay for contrast with white and for matching the green in the top bar.